checkout: integration test: refactor expectations
