+++ test/integrationtest/projecttest/tests/graphql/placeorder_test.go
@@ -16,23 +16,7 @@ import (
 	"github.com/gavv/httpexpect/v2"
 	"github.com/google/go-cmp/cmp"
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 )
-
-// prepareCartWithPaymentSelection adds a simple product via graphQl
-func prepareCartWithPaymentSelection(t *testing.T, e *httpexpect.Expect, paymentMethod string) {
-	t.Helper()
-	helper.GraphQlRequest(t, e, loadGraphQL(t, "add_to_cart", nil)).Expect().Status(http.StatusOK)
-	helper.GraphQlRequest(t, e, loadGraphQL(t, "update_payment_selection", map[string]string{"PAYMENT_METHOD": paymentMethod})).Expect().Status(http.StatusOK)
-}
-
-func updatePaymentSelection(t *testing.T, e *httpexpect.Expect, paymentMethod string) {
-	t.Helper()
-	query := loadGraphQL(t, "update_payment_selection", map[string]string{"PAYMENT_METHOD": paymentMethod})
-
-	response := helper.GraphQlRequest(t, e, query).Expect()
-	response.Status(http.StatusOK)
-}
 
 func Test_PlaceOrderWithPaymentService(t *testing.T) {
 	baseURL := "http://" + FlamingoURL
@@ -173,7 +157,6 @@ func Test_PlaceOrderWithPaymentService(t *testing.T) {
 	}
 }
 
-// TODO: Test place order with fake order service, success / fail / fail during rollback / success during rollback
 func Test_PlaceOrderWithOrderService(t *testing.T) {
 	t.Run("PlaceOrder fails due to payment error, rollback of place order fails, restart afterwards can succeed ", func(t *testing.T) {
 		baseURL := "http://" + FlamingoURL
@@ -195,7 +178,6 @@ func Test_PlaceOrderWithOrderService(t *testing.T) {
 		}
 		helper.AsyncCheckWithTimeout(t, time.Second, func() error {
 			response, refreshUUID := assertRefreshPlaceOrder(t, e, false)
-			require.IsType(t, "string", refreshUUID)
 			assert.Equal(t, uuid, refreshUUID, "uuid has changed")
 			actualState = getValue(response, "Commerce_Checkout_RefreshPlaceOrder", "state").Raw()
 
@@ -215,7 +197,6 @@ func Test_PlaceOrderWithOrderService(t *testing.T) {
 		}
 		helper.AsyncCheckWithTimeout(t, time.Second, func() error {
 			response, refreshUUID := assertRefreshPlaceOrder(t, e, false)
-			require.IsType(t, "string", refreshUUID)
 			assert.Equal(t, uuid, refreshUUID, "uuid has changed")
 			actualState = getValue(response, "Commerce_Checkout_RefreshPlaceOrder", "state").Raw()
 
@@ -242,8 +223,8 @@ func Test_StartPlaceOrder(t *testing.T) {
 
 		actualState := getValue(response, "Commerce_Checkout_RefreshPlaceOrderBlocking", "state")
 		reason := actualState.Object().Value("reason").Object()
-		assert.Equal(t, "Commerce_Checkout_PlaceOrderState_State_FailedReason_PaymentError", reason.Value("__typename").Raw())
-		assert.Equal(t, "PaymentSelection not set", reason.Value("reason").Raw())
+		reason.Value("__typename").Equal("Commerce_Checkout_PlaceOrderState_State_FailedReason_PaymentError")
+		reason.Value("reason").Equal("PaymentSelection not set")
 	})
 
 	t.Run("already running process", func(t *testing.T) {
@@ -294,27 +275,25 @@ func Test_RestartStartPlaceOrder(t *testing.T) {
 	_, uuid1 := assertStartPlaceOrderWithValidUUID(t, e)
 	// wait for fail
 	res, _ := assertRefreshPlaceOrder(t, e, true)
-	state := getValue(res, "Commerce_Checkout_RefreshPlaceOrderBlocking", "state")
-	assert.Equal(t, "Failed", state.Object().Value("name").Raw())
+	getValue(res, "Commerce_Checkout_RefreshPlaceOrderBlocking", "state").Object().Value("name").Equal("Failed")
 
 	// restart
 	_, uuid2 := assertStartPlaceOrderWithValidUUID(t, e)
 	assert.NotEqual(t, uuid1, uuid2, "new process should have been started")
 	res, _ = assertRefreshPlaceOrder(t, e, true)
-	state = getValue(res, "Commerce_Checkout_RefreshPlaceOrderBlocking", "state")
-	assert.Equal(t, "Failed", state.Object().Value("name").Raw())
+	state := getValue(res, "Commerce_Checkout_RefreshPlaceOrderBlocking", "state")
+	state.Object().Value("name").Equal("Failed")
 	reason := state.Object().Value("reason").Object()
 	// payment selection should still be set, so we get the payment error (not PaymentSelection not set)
-	assert.Equal(t, "Commerce_Checkout_PlaceOrderState_State_FailedReason_PaymentError", reason.Value("__typename").String().Raw())
-	assert.Equal(t, "", reason.Value("reason").String().Raw())
+	reason.Value("__typename").Equal("Commerce_Checkout_PlaceOrderState_State_FailedReason_PaymentError")
+	reason.Value("reason").Equal("")
 
 	// update payment selection
 	helper.GraphQlRequest(t, e, loadGraphQL(t, "update_payment_selection", map[string]string{"PAYMENT_METHOD": domain.PaymentFlowStatusCompleted})).Expect().Status(http.StatusOK)
 	_, uuid3 := assertStartPlaceOrderWithValidUUID(t, e)
 	assert.NotEqual(t, uuid2, uuid3, "new process should have been started")
 	res, _ = assertRefreshPlaceOrder(t, e, true)
-	state = getValue(res, "Commerce_Checkout_RefreshPlaceOrderBlocking", "state")
-	assert.Equal(t, "Success", state.Object().Value("name").Raw())
+	getValue(res, "Commerce_Checkout_RefreshPlaceOrderBlocking", "state").Object().Value("name").Equal("Success")
 }
 
 func Test_ActivePlaceOrder(t *testing.T) {
@@ -325,8 +304,7 @@ func Test_ActivePlaceOrder(t *testing.T) {
 	// no process should be running at the start
 	request := helper.GraphQlRequest(t, e, query)
 	response := request.Expect()
-	status := response.JSON().Object().Value("data").Object().Value("Commerce_Checkout_ActivePlaceOrder")
-	assert.False(t, status.Boolean().Raw())
+	response.JSON().Object().Value("data").Object().Value("Commerce_Checkout_ActivePlaceOrder").Boolean().False()
 
 	// let the process wait in iframe status
 	prepareCartWithPaymentSelection(t, e, domain.PaymentFlowActionShowIframe)
@@ -337,8 +315,7 @@ func Test_ActivePlaceOrder(t *testing.T) {
 	// now we have a running process
 	request = helper.GraphQlRequest(t, e, query)
 	response = request.Expect()
-	status = response.JSON().Object().Value("data").Object().Value("Commerce_Checkout_ActivePlaceOrder")
-	assert.True(t, status.Boolean().Raw())
+	response.JSON().Object().Value("data").Object().Value("Commerce_Checkout_ActivePlaceOrder").Boolean().True()
 }
 
 func Test_GetCurrentState(t *testing.T) {
