+++ test/integrationtest/projecttest/tests/graphql/testutil_test.go
@@ -9,8 +9,6 @@ import (
 	"testing"
 
 	"github.com/gavv/httpexpect/v2"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 
 	"flamingo.me/flamingo-commerce/v3/test/integrationtest/projecttest/helper"
 )
@@ -34,6 +32,21 @@ func loadGraphQL(t *testing.T, name string, replacements map[string]string) stri
 	return replacer.Replace(string(content))
 }
 
+// prepareCartWithPaymentSelection adds a simple product via graphQl
+func prepareCartWithPaymentSelection(t *testing.T, e *httpexpect.Expect, paymentMethod string) {
+	t.Helper()
+	helper.GraphQlRequest(t, e, loadGraphQL(t, "add_to_cart", nil)).Expect().Status(http.StatusOK)
+	helper.GraphQlRequest(t, e, loadGraphQL(t, "update_payment_selection", map[string]string{"PAYMENT_METHOD": paymentMethod})).Expect().Status(http.StatusOK)
+}
+
+func updatePaymentSelection(t *testing.T, e *httpexpect.Expect, paymentMethod string) {
+	t.Helper()
+	query := loadGraphQL(t, "update_payment_selection", map[string]string{"PAYMENT_METHOD": paymentMethod})
+
+	response := helper.GraphQlRequest(t, e, query).Expect()
+	response.Status(http.StatusOK)
+}
+
 func assertRefreshPlaceOrder(t *testing.T, e *httpexpect.Expect, blocking bool) (*httpexpect.Response, string) {
 	t.Helper()
 	mutationName := "Commerce_Checkout_RefreshPlaceOrder"
@@ -45,10 +58,9 @@ func assertRefreshPlaceOrder(t *testing.T, e *httpexpect.Expect, blocking bool) 
 	mutation := loadGraphQL(t, fileName, nil)
 	request := helper.GraphQlRequest(t, e, mutation)
 	response := request.Expect()
-	t.Log(response.Body())
-	refreshUUID := getValue(response, mutationName, "uuid").Raw()
-	require.IsType(t, "string", refreshUUID)
-	return response, refreshUUID.(string)
+	refreshUUID := getValue(response, mutationName, "uuid").String().Raw()
+
+	return response, refreshUUID
 }
 
 func assertStartPlaceOrderWithValidUUID(t *testing.T, e *httpexpect.Expect) (*httpexpect.Response, string) {
@@ -58,10 +70,11 @@ func assertStartPlaceOrderWithValidUUID(t *testing.T, e *httpexpect.Expect) (*ht
 	response := request.Expect()
 	t.Log(response.Body())
 	response.Status(http.StatusOK)
-	uuid := getValue(response, "Commerce_Checkout_StartPlaceOrder", "uuid").Raw()
-	require.IsType(t, "string", uuid)
-	assert.Regexp(t, "(?i)^[0-9A-F]{8}-[0-9A-F]{4}-4[0-9A-F]{3}-[89AB][0-9A-F]{3}-[0-9A-F]{12}$", uuid)
-	return response, uuid.(string)
+	uuidMatches := getValue(response, "Commerce_Checkout_StartPlaceOrder", "uuid").String().
+		Match("(?i)^[0-9A-F]{8}-[0-9A-F]{4}-4[0-9A-F]{3}-[89AB][0-9A-F]{3}-[0-9A-F]{12}$")
+	uuidMatches.Length().Equal(1)
+
+	return response, uuidMatches.Index(0).Raw()
 }
 
 func getValue(response *httpexpect.Response, queryName, key string) *httpexpect.Value {
